#24 has been incorporated already
We need to add some evidence for #28

#32 is also implicit in the text

But the most important, make or break it, is #33 and that is difficult to measure. The only thing I can think of is to measure whether time-to-solution is improved or becomes worse with time, or it starts with being better than the average and ends up worse or the obverse. However, this research question is very tricky and we will have to consider it very carefully